Saisie3D : clipping bug et taille de la ball

+++ src/saisie3D/Data.h
@@ -38,7 +38,7 @@ class cData
         void reset();
 
         //Bounding box, center and diameter of all clouds
-        double m_minX, m_maxX, m_minY, m_maxY, m_minZ, m_maxZ, m_cX, m_cY, m_cZ, m_diam;
+        float m_minX, m_maxX, m_minY, m_maxY, m_minZ, m_maxZ, m_cX, m_cY, m_cZ, m_diam;
 
    private:
